requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
apipkg==2.1.1
Appium-Python-Client==2.2.0
async-generator==1.10
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
certifi==2022.5.18
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==37.0.2
cx-Oracle==8.3.0
Deprecated==1.2.13
docopt==0.6.2
execnet==1.9.0
Faker==18.11.2
h11==0.13.0
# idna==3.4
iniconfig==1.1.1
loguru==0.6.0
numpy==1.22.3
opencv-python==4.5.5.64
outcome==1.1.0
packaging==21.3
Pillow==9.1.1
pipreqs==0.4.10
pluggy==1.0.0
py==1.9.0
pycparser==2.21
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyperclip==1.8.1
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
pytesseract==0.3.6
pytest==7.1.2
pytest-forked==1.3.0
pytest-ordering==0.6
pytest-parallel==0.1.1
pytest-rerunfailures==11.1.2
pytest-xdist==2.5.0
python-dateutil==2.8.1
pytweening==1.0.4
PyYAML==6.0
redis==3.5.3
redis-py-cluster==2.1.3
requests==2.27.1
selenium==4.3.0
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
tblib==1.7.0
tesseract==0.1.3
text-unidecode==1.3
toml==0.10.1
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
wrapt==1.14.1
wsproto==1.1.0
xlrd==2.0.1
yarg==0.1.9
urllib3==1.24
loguru==0.6.0
sshtunnel==0.4.0
selenium_stealth==1.0.6